# firmware/docs/openapi.yml
openapi: 3.1.0
info:
  title: Project Verdify MVP API
  version: "2.0"
  description: >
    REST API for AI-powered greenhouse MVP. HTTPS required for all endpoints. Metric units on wire.
    ETag used for config/plan. Device identity via device_name (verdify-aabbcc).
    Rate limiting applies to telemetry endpoints to prevent DoS attacks.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.verdify.ai/api/v1
...
security:
  - UserJWT: []
paths:
...
  /controllers/by-name/{device_name}/config:
    get:
      operationId: getConfigByDeviceName
      tags: [Config]
      summary: Controller fetches materialized config by device_name (ETag supported, gzip compression available)
      security:
        - DeviceToken: []
      parameters:
        - $ref: '#/components/parameters/DeviceNameParam'
        - $ref: '#/components/parameters/IfNoneMatch'
        - name: Accept-Encoding
          in: header
          description: Compression preferences
          schema:
            type: string
            enum: [gzip, identity]
            default: identity
          example: gzip
      responses:
        '200':
          description: Config payload
          headers:
            ETag:
              description: "Strong ETag format. Format: config:v<version>:<sha8> (first 8 hex of SHA-256)"
              schema: { type: string, pattern: '^config:v[0-9]+:[0-9a-f]{8}$' }
            Last-Modified:
              schema: { type: string, format: date-time }
            Content-Encoding:
              description: Compression algorithm used (if any)
              schema:
                type: string
                enum: [gzip, identity]
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigPayload' }
        '304':
          description: Not Modified
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /controllers/me/config:
    get:
      operationId: getMyConfig
      tags: [Config]
      summary: Controller fetches current config for authenticated device (identity inferred from token)
      security:
        - DeviceToken: []
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Configuration payload
          headers:
            ETag:
              description: Strong config ETag config:v<version>:<sha8>
              schema:
                type: string
                pattern: '^config:v[0-9]+:[0-9a-f]{8}$'
            Last-Modified:
              schema: { type: string, format: date-time }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigPayload' }
        '304': { description: Not Modified }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /greenhouses/{id}/config/publish:
    post:
      operationId: publishGreenhouseConfig
      tags: [Config]
      summary: Materialize and publish config snapshot for a greenhouse
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/GreenhouseIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                dry_run: { type: boolean, default: false }
      responses:
        '201':
          description: Published config (new snapshot created)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigPublishResult' }
        '200':
          description: Dry-run preview (no snapshot persisted)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigPublishResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /greenhouses/{id}/config/diff:
    get:
      operationId: getGreenhouseConfigDiff
      tags: [Config]
      summary: Diff current DB config vs last published snapshot
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/GreenhouseIdParam'
      responses:
        '200':
          description: JSON patch-like diff
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigDiff' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # Telemetry ingest (HTTP only)
...
components:
  securitySchemes:
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT for app/admin calls
    DeviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Long-lived device token for controllers
  parameters:
    IfNoneMatch:
      in: header
      name: If-None-Match
      schema: { type: string }
      required: false
    GreenhouseIdParam:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    DeviceNameParam:
      in: path
      name: device_name
      required: true
      schema:
        type: string
        pattern: '^verdify-[0-9a-f]{6}$'
...
  responses:
    BadRequest:
      description: Bad request
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E400_BAD_REQUEST
                message: "Invalid input"
                timestamp: "2025-08-13T18:05:00Z"
                details:
                  field: "email"
                  reason: "Invalid email format"
                request_id: "req_1234567890abcdef"
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E401_UNAUTHORIZED
                message: "Missing/invalid token"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890abcdef"
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E404_NOT_FOUND
                message: "Resource not found"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890"
    Conflict:
      description: Conflict (uniqueness violated)
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E409_CONFLICT
                message: "Unique constraint violated"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890"
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - E400_BAD_REQUEST
        - E401_UNAUTHORIZED
        - E403_FORBIDDEN
        - E404_NOT_FOUND
        - E409_CONFLICT
        - E412_PRECONDITION_FAILED
        - E422_UNPROCESSABLE
        - E429_TOO_MANY_REQUESTS
        - E500_INTERNAL
        - E503_SERVICE_UNAVAILABLE
      description: Standardized error codes used across the API
...
    SensorKind:
      type: string
      enum:
        - temperature
        - humidity
        - vpd
        - co2
        - light
        - soil_moisture
        - water_flow
        - water_total
        - dew_point
        - absolute_humidity
        - enthalpy_delta
        - air_pressure
        - kwh
        - gas_consumption
        - ppfd
        - wind_speed
        - rainfall
        - power
    SensorScope:
      type: string
      enum: [zone, greenhouse, external]
    ActuatorKind:
      type: string
      enum: [fan, heater, vent, fogger, irrigation_valve, fertilizer_valve, pump, light]
...
    ButtonKind:
      type: string
      enum: [cool, heat, humid]

    # --- Error response ---
    ErrorResponse:
      type: object
      required: [error_code, message, timestamp, request_id]
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
          description: Standardized error code
          example: "E409_CONFLICT"
        message:
          type: string
          description: Human-readable error message
          example: "Email already exists"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the error occurred
        request_id:
          type: string
          description: Unique request identifier for tracing
          example: "req_1234567890"
        details:
          oneOf:
            - type: array
              items: { $ref: '#/components/schemas/ErrorDetail' }
            - type: object
              additionalProperties: true
            - type: string
          description: Structured error details for field-specific issues
          example:
            field: "email"
            value: "user@example.com"
...
    ControllerButton:
      type: object
      required: [id, controller_id, button_kind, timeout_s]
      properties:
        id: { type: string, format: uuid }
        controller_id: { type: string, format: uuid }
        button_kind: { $ref: '#/components/schemas/ButtonKind' }
        target_temp_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        target_humi_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        timeout_s: { type: integer, minimum: 1 }
...
    ConfigPayload:
      type: object
      description: Complete configuration package for controllers, including greenhouse parameters, sensor/actuator mappings, and control rules
      required: [version, greenhouse, controllers, sensors, actuators, state_rules, baselines, rails]
      properties:
        version:
          type: integer
          minimum: 1
          description: Configuration version number for ETag generation and change tracking
        generated_at:
          type: string
          format: date-time
          description: UTC timestamp when configuration was generated
        greenhouse:
          type: object
          description: Greenhouse physical parameters and operational limits
          required: [id, title, min_temp_c, max_temp_c, min_vpd_kpa, max_vpd_kpa, enthalpy_open_kjkg, enthalpy_close_kjkg, site_pressure_hpa]
          properties:
            id:
              type: string
              format: uuid
              description: Unique greenhouse identifier
            title:
              type: string
              description: Human-readable greenhouse name
            min_temp_c:
              type: number
              description: Minimum allowable temperature in Celsius (hard safety limit)
            max_temp_c:
              type: number
              description: Maximum allowable temperature in Celsius (hard safety limit)
            min_vpd_kpa:
              type: number
              description: Minimum vapor pressure deficit in kPa (humidity control bound)
            max_vpd_kpa:
              type: number
              description: Maximum vapor pressure deficit in kPa (humidity control bound)
            enthalpy_open_kjkg:
              type: number
              description: Enthalpy threshold in kJ/kg below which ventilation is beneficial for dehumidification
            enthalpy_close_kjkg:
              type: number
              description: Enthalpy threshold in kJ/kg above which ventilation should be reduced
            site_pressure_hpa:
              type: number
              description: Local atmospheric pressure in hPa for enthalpy calculations
        baselines:
          type: object
          description: Control thresholds and hysteresis values for staging decisions
          required: [temp_thresholds, vpd_thresholds, hysteresis]
          properties:
            temp_thresholds:
              type: object
              description: Temperature thresholds in Celsius for staging levels (-3 to +3, where negative indicates heat demand and positive indicates cooling demand)
              required: [minus3, minus2, minus1, zero, plus1, plus2, plus3]
              properties:
                minus3:
                  type: number
                  description: Stage -3 temperature threshold (maximum heat demand)
                minus2:
                  type: number
                  description: Stage -2 temperature threshold (high heat demand)
                minus1:
                  type: number
                  description: Stage -1 temperature threshold (moderate heat demand)
                zero:
                  type: number
                  description: Stage 0 temperature threshold (target/neutral temperature)
                plus1:
                  type: number
                  description: Stage +1 temperature threshold (moderate cooling demand)
                plus2:
                  type: number
                  description: Stage +2 temperature threshold (high cooling demand)
                plus3:
                  type: number
                  description: Stage +3 temperature threshold (maximum cooling demand)
            vpd_thresholds:
              type: object
              description: VPD thresholds in kPa for humidity staging levels (-3 to +3, where negative indicates humidification need and positive indicates dehumidification need)
              required: [minus3, minus2, minus1, zero, plus1, plus2, plus3]
              properties:
                minus3: { type: number }
                minus2: { type: number }
                minus1: { type: number }
                zero: { type: number }
                plus1: { type: number }
                plus2: { type: number }
                plus3: { type: number }
            hysteresis:
              type: object
              properties:
                temp_c: { type: number, default: 0.5 }
                vpd_kpa: { type: number, default: 0.1 }
        rails:
          type: object
          required: [min_temp_c, max_temp_c, min_vpd_kpa, max_vpd_kpa]
          properties:
            min_temp_c: { type: number }
            max_temp_c: { type: number }
            min_vpd_kpa: { type: number }
            max_vpd_kpa: { type: number }
        controllers:
          type: array
          items:
            type: object
            required: [controller_id, device_name, is_climate_controller]
            properties:
              controller_id: { type: string, format: uuid }
              device_name:
                type: string
                pattern: '^verdify-[0-9a-f]{6}$'
              is_climate_controller: { type: boolean }
        sensors:
          type: array
          items:
            type: object
            required: [sensor_id, controller_id, name, kind, scope, include_in_climate_loop]
            properties:
              sensor_id: { type: string, format: uuid }
              controller_id: { type: string, format: uuid }
              name: { type: string }
              kind: { $ref: '#/components/schemas/SensorKind' }
              scope: { $ref: '#/components/schemas/SensorScope' }
              include_in_climate_loop: { type: boolean }
              zone_ids:
                type: array
                items: { type: string, format: uuid }
              poll_interval_s: { type: integer }
        actuators:
          type: array
          items:
            type: object
            required: [actuator_id, controller_id, name, kind]
            properties:
              actuator_id: { type: string, format: uuid }
              controller_id: { type: string, format: uuid }
              name: { type: string }
              kind: { $ref: '#/components/schemas/ActuatorKind' }
              relay_channel: { type: [integer, "null"] }
              min_on_ms: { type: integer }
              min_off_ms: { type: integer }
              zone_id: { type: [string, "null"], format: uuid }
        fan_groups:
          type: array
          items:
            type: object
            required: [fan_group_id, controller_id, name, members]
            properties:
              fan_group_id: { type: string, format: uuid }
              controller_id: { type: string, format: uuid }
              name: { type: string }
              members:
                type: array
                items:
                  type: object
                  required: [actuator_id]
                  properties:
                    actuator_id: { type: string, format: uuid }
        buttons:
          type: array
          items: { $ref: '#/components/schemas/ControllerButton' }
        state_rules:
          type: object
          required: [grid, fallback]
          properties:
            grid:
              type: array
              items:
                type: object
                required: [temp_stage, humi_stage, must_on_actuators, must_off_actuators]
                properties:
                  temp_stage: { type: integer, minimum: -3, maximum: 3 }
                  humi_stage: { type: integer, minimum: -3, maximum: 3 }
                  must_on_actuators:
                    type: array
                    items: { type: string, format: uuid }
                  must_off_actuators:
                    type: array
                    items: { type: string, format: uuid }
                  must_on_fan_groups:
                    type: array
                    items:
                      type: object
                      required: [fan_group_id, on_count]
                      properties:
                        fan_group_id: { type: string, format: uuid }
                        on_count: { type: integer, minimum: 0 }
            fallback:
              type: object
              required: [must_on_actuators, must_off_actuators, must_on_fan_groups, is_fallback]
              properties:
                is_fallback: { type: boolean, enum: [true] }
                must_on_actuators:
                  type: array
                  items: { type: string, format: uuid }
                must_off_actuators:
                  type: array
                  items: { type: string, format: uuid }
                must_on_fan_groups:
                  type: array
                  items:
                    type: object
                    required: [fan_group_id, on_count]
                    properties:
                      fan_group_id: { type: string, format: uuid }
                      on_count: { type: integer, minimum: 0 }
    ConfigPublishResult:
      type: object
      required: [published, version, errors, warnings, payload]
      properties:
        published: { type: boolean }
        version: { type: integer }
        errors:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
        payload:
          $ref: '#/components/schemas/ConfigPayload'
    ConfigDiff:
      type: object
      properties:
        added: { type: array, items: { type: string } }
        removed: { type: array, items: { type: string } }
        changed: { type: array, items: { type: string } }

    # --- Plan schemas ---
...